Move more constructors, astype, where
