+++ python/cudf/cudf/tests/test_dataframe.py
@@ -6,12 +6,11 @@ import decimal
 import functools
 import io
 import operator
-import random
 import re
 import string
 import textwrap
 import warnings
-from collections import OrderedDict, defaultdict, namedtuple
+from collections import OrderedDict, namedtuple
 from contextlib import contextmanager
 from copy import copy
 
@@ -20,7 +19,6 @@ import numpy as np
 import pandas as pd
 import pyarrow as pa
 import pytest
-from numba import cuda
 from packaging import version
 
 import cudf
@@ -3227,45 +3225,6 @@ def test_tupleize_cols_False_set():
     assert_eq(pdf.columns, gdf.columns)
 
 
-def test_init_multiindex_from_dict():
-    pdf = pd.DataFrame({("a", "b"): [1]})
-    gdf = cudf.DataFrame({("a", "b"): [1]})
-    assert_eq(pdf, gdf)
-    assert_eq(pdf.columns, gdf.columns)
-
-
-def test_change_column_dtype_in_empty():
-    pdf = pd.DataFrame({"a": [], "b": []})
-    gdf = cudf.from_pandas(pdf)
-    assert_eq(pdf, gdf)
-    pdf["b"] = pdf["b"].astype("int64")
-    gdf["b"] = gdf["b"].astype("int64")
-    assert_eq(pdf, gdf)
-
-
-@pytest.mark.parametrize("dtype", ["int64", "str"])
-def test_dataframe_from_dictionary_series_same_name_index(dtype):
-    pd_idx1 = pd.Index([1, 2, 0], name="test_index").astype(dtype)
-    pd_idx2 = pd.Index([2, 0, 1], name="test_index").astype(dtype)
-    pd_series1 = pd.Series([1, 2, 3], index=pd_idx1)
-    pd_series2 = pd.Series([1, 2, 3], index=pd_idx2)
-
-    gd_idx1 = cudf.from_pandas(pd_idx1)
-    gd_idx2 = cudf.from_pandas(pd_idx2)
-    gd_series1 = cudf.Series([1, 2, 3], index=gd_idx1)
-    gd_series2 = cudf.Series([1, 2, 3], index=gd_idx2)
-
-    expect = pd.DataFrame({"a": pd_series1, "b": pd_series2})
-    got = cudf.DataFrame({"a": gd_series1, "b": gd_series2})
-
-    if dtype == "str":
-        # Pandas actually loses its index name erroneously here...
-        expect.index.name = "test_index"
-
-    assert_eq(expect, got)
-    assert expect.index.names == got.index.names
-
-
 @pytest.mark.parametrize(
     "arg", [slice(2, 8, 3), slice(1, 20, 4), slice(-2, -6, -2)]
 )
@@ -3286,729 +3245,6 @@ def test_dataframe_strided_slice(arg):
     got = gdf[arg]
 
     assert_eq(expect, got)
-
-
-@pytest.mark.parametrize(
-    "data,condition,other,error",
-    [
-        (pd.Series(range(5)), pd.Series(range(5)) > 0, None, None),
-        (pd.Series(range(5)), pd.Series(range(5)) > 1, None, None),
-        (pd.Series(range(5)), pd.Series(range(5)) > 1, 10, None),
-        (
-            pd.Series(range(5)),
-            pd.Series(range(5)) > 1,
-            pd.Series(range(5, 10)),
-            None,
-        ),
-        (
-            pd.DataFrame(np.arange(10).reshape(-1, 2), columns=["A", "B"]),
-            (
-                pd.DataFrame(np.arange(10).reshape(-1, 2), columns=["A", "B"])
-                % 3
-            )
-            == 0,
-            -pd.DataFrame(np.arange(10).reshape(-1, 2), columns=["A", "B"]),
-            None,
-        ),
-        (
-            pd.DataFrame({"a": [1, 2, np.nan], "b": [4, np.nan, 6]}),
-            pd.DataFrame({"a": [1, 2, np.nan], "b": [4, np.nan, 6]}) == 4,
-            None,
-            None,
-        ),
-        (
-            pd.DataFrame({"a": [1, 2, np.nan], "b": [4, np.nan, 6]}),
-            pd.DataFrame({"a": [1, 2, np.nan], "b": [4, np.nan, 6]}) != 4,
-            None,
-            None,
-        ),
-        (
-            pd.DataFrame({"p": [-2, 3, -4, -79], "k": [9, 10, 11, 12]}),
-            [True, True, True],
-            None,
-            ValueError,
-        ),
-        (
-            pd.DataFrame({"p": [-2, 3, -4, -79], "k": [9, 10, 11, 12]}),
-            [True, True, True, False],
-            None,
-            ValueError,
-        ),
-        (
-            pd.DataFrame({"p": [-2, 3, -4, -79], "k": [9, 10, 11, 12]}),
-            [[True, True, True, False], [True, True, True, False]],
-            None,
-            ValueError,
-        ),
-        (
-            pd.DataFrame({"p": [-2, 3, -4, -79], "k": [9, 10, 11, 12]}),
-            [[True, True], [False, True], [True, False], [False, True]],
-            None,
-            None,
-        ),
-        (
-            pd.DataFrame({"p": [-2, 3, -4, -79], "k": [9, 10, 11, 12]}),
-            cuda.to_device(
-                np.array(
-                    [[True, True], [False, True], [True, False], [False, True]]
-                )
-            ),
-            None,
-            None,
-        ),
-        (
-            pd.DataFrame({"p": [-2, 3, -4, -79], "k": [9, 10, 11, 12]}),
-            cupy.array(
-                [[True, True], [False, True], [True, False], [False, True]]
-            ),
-            17,
-            None,
-        ),
-        (
-            pd.DataFrame({"p": [-2, 3, -4, -79], "k": [9, 10, 11, 12]}),
-            [[True, True], [False, True], [True, False], [False, True]],
-            17,
-            None,
-        ),
-        (
-            pd.DataFrame({"p": [-2, 3, -4, -79], "k": [9, 10, 11, 12]}),
-            [
-                [True, True, False, True],
-                [True, True, False, True],
-                [True, True, False, True],
-                [True, True, False, True],
-            ],
-            None,
-            ValueError,
-        ),
-        (
-            pd.Series([1, 2, np.nan]),
-            pd.Series([1, 2, np.nan]) == 4,
-            None,
-            None,
-        ),
-        (
-            pd.Series([1, 2, np.nan]),
-            pd.Series([1, 2, np.nan]) != 4,
-            None,
-            None,
-        ),
-        (
-            pd.Series([4, np.nan, 6]),
-            pd.Series([4, np.nan, 6]) == 4,
-            None,
-            None,
-        ),
-        (
-            pd.Series([4, np.nan, 6]),
-            pd.Series([4, np.nan, 6]) != 4,
-            None,
-            None,
-        ),
-        (
-            pd.Series([4, np.nan, 6], dtype="category"),
-            pd.Series([4, np.nan, 6], dtype="category") != 4,
-            None,
-            None,
-        ),
-        (
-            pd.Series(["a", "b", "b", "d", "c", "s"], dtype="category"),
-            pd.Series(["a", "b", "b", "d", "c", "s"], dtype="category") == "b",
-            None,
-            None,
-        ),
-        (
-            pd.Series(["a", "b", "b", "d", "c", "s"], dtype="category"),
-            pd.Series(["a", "b", "b", "d", "c", "s"], dtype="category") == "b",
-            "s",
-            None,
-        ),
-        (
-            pd.Series([1, 2, 3, 2, 5]),
-            pd.Series([1, 2, 3, 2, 5]) == 2,
-            pd.DataFrame(
-                {
-                    "a": pd.Series([1, 2, 3, 2, 5]),
-                    "b": pd.Series([1, 2, 3, 2, 5]),
-                }
-            ),
-            NotImplementedError,
-        ),
-    ],
-)
-@pytest.mark.parametrize("inplace", [True, False])
-def test_df_sr_mask_where(data, condition, other, error, inplace):
-    ps_where = data
-    gs_where = cudf.from_pandas(data)
-
-    ps_mask = ps_where.copy(deep=True)
-    gs_mask = gs_where.copy(deep=True)
-
-    if hasattr(condition, "__cuda_array_interface__"):
-        if type(condition).__module__.split(".")[0] == "cupy":
-            ps_condition = cupy.asnumpy(condition)
-        else:
-            ps_condition = np.array(condition).astype("bool")
-    else:
-        ps_condition = condition
-
-    if type(condition).__module__.split(".")[0] == "pandas":
-        gs_condition = cudf.from_pandas(condition)
-    else:
-        gs_condition = condition
-
-    ps_other = other
-    if type(other).__module__.split(".")[0] == "pandas":
-        gs_other = cudf.from_pandas(other)
-    else:
-        gs_other = other
-
-    if error is None:
-        expect_where = ps_where.where(
-            ps_condition, other=ps_other, inplace=inplace
-        )
-        got_where = gs_where.where(
-            gs_condition, other=gs_other, inplace=inplace
-        )
-
-        expect_mask = ps_mask.mask(
-            ps_condition, other=ps_other, inplace=inplace
-        )
-        got_mask = gs_mask.mask(gs_condition, other=gs_other, inplace=inplace)
-
-        if inplace:
-            expect_where = ps_where
-            got_where = gs_where
-
-            expect_mask = ps_mask
-            got_mask = gs_mask
-
-        if isinstance(expect_where, pd.Series) and isinstance(
-            expect_where.dtype, pd.CategoricalDtype
-        ):
-            np.testing.assert_array_equal(
-                expect_where.cat.codes,
-                got_where.cat.codes.astype(expect_where.cat.codes.dtype)
-                .fillna(-1)
-                .to_numpy(),
-            )
-            assert_eq(expect_where.cat.categories, got_where.cat.categories)
-
-            np.testing.assert_array_equal(
-                expect_mask.cat.codes,
-                got_mask.cat.codes.astype(expect_mask.cat.codes.dtype)
-                .fillna(-1)
-                .to_numpy(),
-            )
-            assert_eq(expect_mask.cat.categories, got_mask.cat.categories)
-        else:
-            assert_eq(
-                expect_where.fillna(-1),
-                got_where.fillna(-1),
-                check_dtype=False,
-            )
-            assert_eq(
-                expect_mask.fillna(-1), got_mask.fillna(-1), check_dtype=False
-            )
-    else:
-        assert_exceptions_equal(
-            lfunc=ps_where.where,
-            rfunc=gs_where.where,
-            lfunc_args_and_kwargs=(
-                [ps_condition],
-                {"other": ps_other, "inplace": inplace},
-            ),
-            rfunc_args_and_kwargs=(
-                [gs_condition],
-                {"other": gs_other, "inplace": inplace},
-            ),
-        )
-
-        assert_exceptions_equal(
-            lfunc=ps_mask.mask,
-            rfunc=gs_mask.mask,
-            lfunc_args_and_kwargs=(
-                [ps_condition],
-                {"other": ps_other, "inplace": inplace},
-            ),
-            rfunc_args_and_kwargs=(
-                [gs_condition],
-                {"other": gs_other, "inplace": inplace},
-            ),
-        )
-
-
-@pytest.mark.parametrize(
-    "data,condition,other,has_cat",
-    [
-        (
-            pd.DataFrame(
-                {
-                    "a": pd.Series(["a", "a", "b", "c", "a", "d", "d", "a"]),
-                    "b": pd.Series(["o", "p", "q", "e", "p", "p", "a", "a"]),
-                }
-            ),
-            pd.DataFrame(
-                {
-                    "a": pd.Series(["a", "a", "b", "c", "a", "d", "d", "a"]),
-                    "b": pd.Series(["o", "p", "q", "e", "p", "p", "a", "a"]),
-                }
-            )
-            != "a",
-            None,
-            None,
-        ),
-        (
-            pd.DataFrame(
-                {
-                    "a": pd.Series(
-                        ["a", "a", "b", "c", "a", "d", "d", "a"],
-                        dtype="category",
-                    ),
-                    "b": pd.Series(
-                        ["o", "p", "q", "e", "p", "p", "a", "a"],
-                        dtype="category",
-                    ),
-                }
-            ),
-            pd.DataFrame(
-                {
-                    "a": pd.Series(
-                        ["a", "a", "b", "c", "a", "d", "d", "a"],
-                        dtype="category",
-                    ),
-                    "b": pd.Series(
-                        ["o", "p", "q", "e", "p", "p", "a", "a"],
-                        dtype="category",
-                    ),
-                }
-            )
-            != "a",
-            None,
-            True,
-        ),
-        (
-            pd.DataFrame(
-                {
-                    "a": pd.Series(
-                        ["a", "a", "b", "c", "a", "d", "d", "a"],
-                        dtype="category",
-                    ),
-                    "b": pd.Series(
-                        ["o", "p", "q", "e", "p", "p", "a", "a"],
-                        dtype="category",
-                    ),
-                }
-            ),
-            pd.DataFrame(
-                {
-                    "a": pd.Series(
-                        ["a", "a", "b", "c", "a", "d", "d", "a"],
-                        dtype="category",
-                    ),
-                    "b": pd.Series(
-                        ["o", "p", "q", "e", "p", "p", "a", "a"],
-                        dtype="category",
-                    ),
-                }
-            )
-            == "a",
-            None,
-            True,
-        ),
-        (
-            pd.DataFrame(
-                {
-                    "a": pd.Series(
-                        ["a", "a", "b", "c", "a", "d", "d", "a"],
-                        dtype="category",
-                    ),
-                    "b": pd.Series(
-                        ["o", "p", "q", "e", "p", "p", "a", "a"],
-                        dtype="category",
-                    ),
-                }
-            ),
-            pd.DataFrame(
-                {
-                    "a": pd.Series(
-                        ["a", "a", "b", "c", "a", "d", "d", "a"],
-                        dtype="category",
-                    ),
-                    "b": pd.Series(
-                        ["o", "p", "q", "e", "p", "p", "a", "a"],
-                        dtype="category",
-                    ),
-                }
-            )
-            != "a",
-            "a",
-            True,
-        ),
-        (
-            pd.DataFrame(
-                {
-                    "a": pd.Series(
-                        ["a", "a", "b", "c", "a", "d", "d", "a"],
-                        dtype="category",
-                    ),
-                    "b": pd.Series(
-                        ["o", "p", "q", "e", "p", "p", "a", "a"],
-                        dtype="category",
-                    ),
-                }
-            ),
-            pd.DataFrame(
-                {
-                    "a": pd.Series(
-                        ["a", "a", "b", "c", "a", "d", "d", "a"],
-                        dtype="category",
-                    ),
-                    "b": pd.Series(
-                        ["o", "p", "q", "e", "p", "p", "a", "a"],
-                        dtype="category",
-                    ),
-                }
-            )
-            == "a",
-            "a",
-            True,
-        ),
-    ],
-)
-def test_df_string_cat_types_mask_where(data, condition, other, has_cat):
-    ps = data
-    gs = cudf.from_pandas(data)
-
-    ps_condition = condition
-    if type(condition).__module__.split(".")[0] == "pandas":
-        gs_condition = cudf.from_pandas(condition)
-    else:
-        gs_condition = condition
-
-    ps_other = other
-    if type(other).__module__.split(".")[0] == "pandas":
-        gs_other = cudf.from_pandas(other)
-    else:
-        gs_other = other
-
-    expect_where = ps.where(ps_condition, other=ps_other)
-    got_where = gs.where(gs_condition, other=gs_other)
-
-    expect_mask = ps.mask(ps_condition, other=ps_other)
-    got_mask = gs.mask(gs_condition, other=gs_other)
-
-    if has_cat is None:
-        assert_eq(
-            expect_where.fillna(-1).astype("str"),
-            got_where.fillna(-1),
-            check_dtype=False,
-        )
-        assert_eq(
-            expect_mask.fillna(-1).astype("str"),
-            got_mask.fillna(-1),
-            check_dtype=False,
-        )
-    else:
-        assert_eq(expect_where, got_where, check_dtype=False)
-        assert_eq(expect_mask, got_mask, check_dtype=False)
-
-
-@pytest.mark.parametrize(
-    "data,expected_upcast_type,error",
-    [
-        (
-            pd.Series([random.random() for _ in range(10)], dtype="float32"),
-            np.dtype("float32"),
-            None,
-        ),
-        (
-            pd.Series([random.random() for _ in range(10)], dtype="float16"),
-            None,
-            TypeError,
-        ),
-        (
-            pd.Series([random.random() for _ in range(10)], dtype="float64"),
-            np.dtype("float64"),
-            None,
-        ),
-        (
-            pd.Series([random.random() for _ in range(10)], dtype="float128"),
-            None,
-            ValueError,
-        ),
-    ],
-)
-def test_from_pandas_unsupported_types(data, expected_upcast_type, error):
-    pdf = pd.DataFrame({"one_col": data})
-    if error is not None:
-        with pytest.raises(error):
-            cudf.from_pandas(data)
-
-        with pytest.raises(error):
-            cudf.Series(data)
-
-        with pytest.raises(error):
-            cudf.from_pandas(pdf)
-
-        with pytest.raises(error):
-            cudf.DataFrame(pdf)
-    else:
-        df = cudf.from_pandas(data)
-
-        assert_eq(data, df, check_dtype=False)
-        assert df.dtype == expected_upcast_type
-
-        df = cudf.Series(data)
-        assert_eq(data, df, check_dtype=False)
-        assert df.dtype == expected_upcast_type
-
-        df = cudf.from_pandas(pdf)
-        assert_eq(pdf, df, check_dtype=False)
-        assert df["one_col"].dtype == expected_upcast_type
-
-        df = cudf.DataFrame(pdf)
-        assert_eq(pdf, df, check_dtype=False)
-        assert df["one_col"].dtype == expected_upcast_type
-
-
-@pytest.mark.parametrize("nan_as_null", [True, False])
-@pytest.mark.parametrize("index", [None, "a", ["a", "b"]])
-def test_from_pandas_nan_as_null(nan_as_null, index):
-    data = [np.nan, 2.0, 3.0]
-
-    if index is None:
-        pdf = pd.DataFrame({"a": data, "b": data})
-        expected = cudf.DataFrame(
-            {
-                "a": as_column(data, nan_as_null=nan_as_null),
-                "b": as_column(data, nan_as_null=nan_as_null),
-            }
-        )
-    else:
-        pdf = pd.DataFrame({"a": data, "b": data}).set_index(index)
-        expected = cudf.DataFrame(
-            {
-                "a": as_column(data, nan_as_null=nan_as_null),
-                "b": as_column(data, nan_as_null=nan_as_null),
-            }
-        )
-        expected = cudf.DataFrame(
-            {
-                "a": as_column(data, nan_as_null=nan_as_null),
-                "b": as_column(data, nan_as_null=nan_as_null),
-            }
-        )
-        expected = expected.set_index(index)
-
-    got = cudf.from_pandas(pdf, nan_as_null=nan_as_null)
-
-    assert_eq(expected, got)
-
-
-@pytest.mark.parametrize("nan_as_null", [True, False])
-def test_from_pandas_for_series_nan_as_null(nan_as_null):
-    data = [np.nan, 2.0, 3.0]
-    psr = pd.Series(data)
-
-    expected = cudf.Series._from_column(
-        as_column(data, nan_as_null=nan_as_null)
-    )
-    got = cudf.from_pandas(psr, nan_as_null=nan_as_null)
-
-    assert_eq(expected, got)
-
-
-@pytest.mark.parametrize("copy", [True, False])
-def test_df_series_dataframe_astype_copy(copy):
-    gdf = cudf.DataFrame({"col1": [1, 2], "col2": [3, 4]})
-    pdf = gdf.to_pandas()
-
-    assert_eq(
-        gdf.astype(dtype="float", copy=copy),
-        pdf.astype(dtype="float", copy=copy),
-    )
-    assert_eq(gdf, pdf)
-
-    gsr = cudf.Series([1, 2])
-    psr = gsr.to_pandas()
-
-    assert_eq(
-        gsr.astype(dtype="float", copy=copy),
-        psr.astype(dtype="float", copy=copy),
-    )
-    assert_eq(gsr, psr)
-
-    gsr = cudf.Series([1, 2])
-    psr = gsr.to_pandas()
-
-    actual = gsr.astype(dtype="int64", copy=copy)
-    expected = psr.astype(dtype="int64", copy=copy)
-    assert_eq(expected, actual)
-    assert_eq(gsr, psr)
-    actual[0] = 3
-    expected[0] = 3
-    assert_eq(gsr, psr)
-
-
-@pytest.mark.parametrize("copy", [True, False])
-def test_df_series_dataframe_astype_dtype_dict(copy):
-    gdf = cudf.DataFrame({"col1": [1, 2], "col2": [3, 4]})
-    pdf = gdf.to_pandas()
-
-    assert_eq(
-        gdf.astype(dtype={"col1": "float"}, copy=copy),
-        pdf.astype(dtype={"col1": "float"}, copy=copy),
-    )
-    assert_eq(gdf, pdf)
-
-    gsr = cudf.Series([1, 2])
-    psr = gsr.to_pandas()
-
-    assert_eq(
-        gsr.astype(dtype={None: "float"}, copy=copy),
-        psr.astype(dtype={None: "float"}, copy=copy),
-    )
-    assert_eq(gsr, psr)
-
-    assert_exceptions_equal(
-        lfunc=psr.astype,
-        rfunc=gsr.astype,
-        lfunc_args_and_kwargs=([], {"dtype": {"a": "float"}, "copy": copy}),
-        rfunc_args_and_kwargs=([], {"dtype": {"a": "float"}, "copy": copy}),
-    )
-
-    gsr = cudf.Series([1, 2])
-    psr = gsr.to_pandas()
-
-    actual = gsr.astype({None: "int64"}, copy=copy)
-    expected = psr.astype({None: "int64"}, copy=copy)
-    assert_eq(expected, actual)
-    assert_eq(gsr, psr)
-
-    actual[0] = 3
-    expected[0] = 3
-    assert_eq(gsr, psr)
-
-
-@pytest.mark.parametrize(
-    "data,columns",
-    [
-        ([1, 2, 3, 100, 112, 35464], ["a"]),
-        (range(100), None),
-        (
-            [],
-            None,
-        ),
-        ((-10, 21, 32, 32, 1, 2, 3), ["p"]),
-        (
-            (),
-            None,
-        ),
-        ([[1, 2, 3], [1, 2, 3]], ["col1", "col2", "col3"]),
-        ([range(100), range(100)], ["range" + str(i) for i in range(100)]),
-        (((1, 2, 3), (1, 2, 3)), ["tuple0", "tuple1", "tuple2"]),
-        ([[1, 2, 3]], ["list col1", "list col2", "list col3"]),
-        ([[1, 2, 3]], pd.Index(["col1", "col2", "col3"], name="rapids")),
-        ([range(100)], ["range" + str(i) for i in range(100)]),
-        (((1, 2, 3),), ["k1", "k2", "k3"]),
-    ],
-)
-def test_dataframe_init_1d_list(data, columns):
-    expect = pd.DataFrame(data, columns=columns)
-    actual = cudf.DataFrame(data, columns=columns)
-
-    assert_eq(
-        expect,
-        actual,
-        check_index_type=len(data) != 0,
-    )
-
-    expect = pd.DataFrame(data, columns=None)
-    actual = cudf.DataFrame(data, columns=None)
-
-    assert_eq(
-        expect,
-        actual,
-        check_index_type=len(data) != 0,
-    )
-
-
-@pytest.mark.parametrize(
-    "data,cols,index",
-    [
-        (
-            np.ndarray(shape=(4, 2), dtype=float, order="F"),
-            ["a", "b"],
-            ["a", "b", "c", "d"],
-        ),
-        (
-            np.ndarray(shape=(4, 2), dtype=float, order="F"),
-            ["a", "b"],
-            [0, 20, 30, 10],
-        ),
-        (
-            np.ndarray(shape=(4, 2), dtype=float, order="F"),
-            ["a", "b"],
-            [0, 1, 2, 3],
-        ),
-        (np.array([11, 123, -2342, 232]), ["a"], [1, 2, 11, 12]),
-        (np.array([11, 123, -2342, 232]), ["a"], ["khsdjk", "a", "z", "kk"]),
-        (
-            cupy.ndarray(shape=(4, 2), dtype=float, order="F"),
-            ["a", "z"],
-            ["a", "z", "a", "z"],
-        ),
-        (cupy.array([11, 123, -2342, 232]), ["z"], [0, 1, 1, 0]),
-        (cupy.array([11, 123, -2342, 232]), ["z"], [1, 2, 3, 4]),
-        (cupy.array([11, 123, -2342, 232]), ["z"], ["a", "z", "d", "e"]),
-        (
-            np.random.default_rng(seed=0).standard_normal(size=(2, 4)),
-            ["a", "b", "c", "d"],
-            ["a", "b"],
-        ),
-        (
-            np.random.default_rng(seed=0).standard_normal(size=(2, 4)),
-            ["a", "b", "c", "d"],
-            [1, 0],
-        ),
-        (cupy.random.randn(2, 4), ["a", "b", "c", "d"], ["a", "b"]),
-        (cupy.random.randn(2, 4), ["a", "b", "c", "d"], [1, 0]),
-    ],
-)
-def test_dataframe_init_from_arrays_cols(data, cols, index):
-    gd_data = data
-    if isinstance(data, cupy.ndarray):
-        # pandas can't handle cupy arrays in general
-        pd_data = data.get()
-
-        # additional test for building DataFrame with gpu array whose
-        # cuda array interface has no `descr` attribute
-        numba_data = cuda.as_cuda_array(data)
-    else:
-        pd_data = data
-        numba_data = None
-
-    # verify with columns & index
-    pdf = pd.DataFrame(pd_data, columns=cols, index=index)
-    gdf = cudf.DataFrame(gd_data, columns=cols, index=index)
-
-    assert_eq(pdf, gdf, check_dtype=False)
-
-    # verify with columns
-    pdf = pd.DataFrame(pd_data, columns=cols)
-    gdf = cudf.DataFrame(gd_data, columns=cols)
-
-    assert_eq(pdf, gdf, check_dtype=False)
-
-    pdf = pd.DataFrame(pd_data)
-    gdf = cudf.DataFrame(gd_data)
-
-    assert_eq(pdf, gdf, check_dtype=False)
-
-    if numba_data is not None:
-        gdf = cudf.DataFrame(numba_data)
-        assert_eq(pdf, gdf, check_dtype=False)
 
 
 @pytest_unmark_spilling
@@ -4360,172 +3596,6 @@ def test_dataframe_info_null_counts():
     df.info(buf=buffer, null_counts=True)
     actual_string = buffer.getvalue()
     assert str_cmp == actual_string
-
-
-@pytest.mark.parametrize(
-    "orient", ["dict", "list", "split", "tight", "records", "index", "series"]
-)
-@pytest.mark.parametrize("into", [dict, OrderedDict, defaultdict(list)])
-def test_dataframe_to_dict(orient, into):
-    df = cudf.DataFrame({"a": [1, 2, 3], "b": [9, 5, 3]}, index=[10, 11, 12])
-    pdf = df.to_pandas()
-
-    actual = df.to_dict(orient=orient, into=into)
-    expected = pdf.to_dict(orient=orient, into=into)
-    if orient == "series":
-        assert actual.keys() == expected.keys()
-        for key in actual.keys():
-            assert_eq(expected[key], actual[key])
-    else:
-        assert_eq(expected, actual)
-
-
-@pytest.mark.parametrize(
-    "data, orient, dtype, columns",
-    [
-        (
-            {"col_1": [3, 2, 1, 0], "col_2": [3, 2, 1, 0]},
-            "columns",
-            None,
-            None,
-        ),
-        ({"col_1": [3, 2, 1, 0], "col_2": [3, 2, 1, 0]}, "index", None, None),
-        (
-            {"col_1": [None, 2, 1, 0], "col_2": [3, None, 1, 0]},
-            "index",
-            None,
-            ["A", "B", "C", "D"],
-        ),
-        (
-            {
-                "col_1": ["ab", "cd", "ef", "gh"],
-                "col_2": ["zx", "one", "two", "three"],
-            },
-            "index",
-            None,
-            ["A", "B", "C", "D"],
-        ),
-        (
-            {
-                "index": [("a", "b"), ("a", "c")],
-                "columns": [("x", 1), ("y", 2)],
-                "data": [[1, 3], [2, 4]],
-                "index_names": ["n1", "n2"],
-                "column_names": ["z1", "z2"],
-            },
-            "tight",
-            "float64",
-            None,
-        ),
-    ],
-)
-def test_dataframe_from_dict(data, orient, dtype, columns):
-    expected = pd.DataFrame.from_dict(
-        data=data, orient=orient, dtype=dtype, columns=columns
-    )
-
-    actual = cudf.DataFrame.from_dict(
-        data=data, orient=orient, dtype=dtype, columns=columns
-    )
-
-    assert_eq(expected, actual)
-
-
-@pytest.mark.parametrize("dtype", ["int64", "str", None])
-def test_dataframe_from_dict_transposed(dtype):
-    pd_data = {"a": [3, 2, 1, 0], "col_2": [3, 2, 1, 0]}
-    gd_data = {key: cudf.Series(val) for key, val in pd_data.items()}
-
-    expected = pd.DataFrame.from_dict(pd_data, orient="index", dtype=dtype)
-    actual = cudf.DataFrame.from_dict(gd_data, orient="index", dtype=dtype)
-
-    gd_data = {key: cupy.asarray(val) for key, val in pd_data.items()}
-    actual = cudf.DataFrame.from_dict(gd_data, orient="index", dtype=dtype)
-    assert_eq(expected, actual)
-
-
-@pytest.mark.parametrize(
-    "pd_data, gd_data, orient, dtype, columns",
-    [
-        (
-            {"col_1": np.array([3, 2, 1, 0]), "col_2": np.array([3, 2, 1, 0])},
-            {
-                "col_1": cupy.array([3, 2, 1, 0]),
-                "col_2": cupy.array([3, 2, 1, 0]),
-            },
-            "columns",
-            None,
-            None,
-        ),
-        (
-            {"col_1": np.array([3, 2, 1, 0]), "col_2": np.array([3, 2, 1, 0])},
-            {
-                "col_1": cupy.array([3, 2, 1, 0]),
-                "col_2": cupy.array([3, 2, 1, 0]),
-            },
-            "index",
-            None,
-            None,
-        ),
-        (
-            {
-                "col_1": np.array([None, 2, 1, 0]),
-                "col_2": np.array([3, None, 1, 0]),
-            },
-            {
-                "col_1": cupy.array([np.nan, 2, 1, 0]),
-                "col_2": cupy.array([3, np.nan, 1, 0]),
-            },
-            "index",
-            None,
-            ["A", "B", "C", "D"],
-        ),
-        (
-            {
-                "col_1": np.array(["ab", "cd", "ef", "gh"]),
-                "col_2": np.array(["zx", "one", "two", "three"]),
-            },
-            {
-                "col_1": np.array(["ab", "cd", "ef", "gh"]),
-                "col_2": np.array(["zx", "one", "two", "three"]),
-            },
-            "index",
-            None,
-            ["A", "B", "C", "D"],
-        ),
-        (
-            {
-                "index": [("a", "b"), ("a", "c")],
-                "columns": [("x", 1), ("y", 2)],
-                "data": [np.array([1, 3]), np.array([2, 4])],
-                "index_names": ["n1", "n2"],
-                "column_names": ["z1", "z2"],
-            },
-            {
-                "index": [("a", "b"), ("a", "c")],
-                "columns": [("x", 1), ("y", 2)],
-                "data": [cupy.array([1, 3]), cupy.array([2, 4])],
-                "index_names": ["n1", "n2"],
-                "column_names": ["z1", "z2"],
-            },
-            "tight",
-            "float64",
-            None,
-        ),
-    ],
-)
-def test_dataframe_from_dict_cp_np_arrays(
-    pd_data, gd_data, orient, dtype, columns
-):
-    expected = pd.DataFrame.from_dict(
-        data=pd_data, orient=orient, dtype=dtype, columns=columns
-    )
-
-    actual = cudf.DataFrame.from_dict(
-        data=gd_data, orient=orient, dtype=dtype, columns=columns
-    )
-
-    assert_eq(expected, actual, check_dtype=dtype is not None)
 
 
 @pytest.mark.parametrize(
@@ -6423,56 +5493,6 @@ def test_dataframe_argsort(data, ascending, expected_data):
     expected = cupy.array(expected_data, dtype="int32")
 
     assert_eq(actual, expected)
-
-
-@pytest.mark.parametrize(
-    "data,columns,index",
-    [
-        (pd.Series([1, 2, 3]), None, None),
-        (pd.Series(["a", "b", None, "c"], name="abc"), None, None),
-        (
-            pd.Series(["a", "b", None, "c"], name="abc"),
-            ["abc", "b"],
-            [1, 2, 3],
-        ),
-    ],
-)
-def test_dataframe_init_from_series(data, columns, index):
-    expected = pd.DataFrame(data, columns=columns, index=index)
-    actual = cudf.DataFrame(data, columns=columns, index=index)
-
-    assert_eq(
-        expected,
-        actual,
-        check_index_type=len(expected) != 0,
-    )
-
-
-def test_frame_series_where():
-    gdf = cudf.DataFrame(
-        {"a": [1.0, 2.0, None, 3.0, None], "b": [None, 10.0, 11.0, None, 23.0]}
-    )
-    pdf = gdf.to_pandas()
-    expected = gdf.where(gdf.notna(), gdf.mean())
-    actual = pdf.where(pdf.notna(), pdf.mean(), axis=1)
-    assert_eq(expected, actual)
-
-
-def test_frame_series_where_other():
-    gdf = cudf.DataFrame({"a": [1, 2, 3], "b": [1, 1, 0]})
-    pdf = gdf.to_pandas()
-
-    expected = gdf.where(gdf["b"] == 1, cudf.NA)
-    actual = pdf.where(pdf["b"] == 1, pd.NA)
-    assert_eq(
-        actual.fillna(-1).values,
-        expected.fillna(-1).values,
-        check_dtype=False,
-    )
-
-    expected = gdf.where(gdf["b"] == 1, 0)
-    actual = pdf.where(pdf["b"] == 1, 0)
-    assert_eq(expected, actual)
 
 
 @pytest.mark.parametrize(
